requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
boto3==1.21.32
botocore==1.24.32
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
crispy-bootstrap5==0.6
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.3
django-ajax==0.3.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-storages==1.12.3
et-xmlfile==1.1.0
filelock==3.6.0
future==0.18.2
gdown==4.4.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
openpyxl==3.0.9
Pillow==9.1.0
psycopg2==2.9.3
PySocks==1.7.1
python-dateutil==2.8.2
ratelim==0.1.6
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tqdm==4.63.1
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
django-filter~=22.1
django-admin-sortable2~=2.1.10
pdf2image~=1.16.2
django-ckeditor~=6.7.0